Splitted utils file
Utils file now is a folder and it has been splitted into different files, each containing functions grouped by its scope of action
