+++ lambda/utils/STATES.py
@@ -1,0 +1,5 @@
+STATES = {
+  'INITIALIZED': 0,
+  'QUESTION_ASKED': 1,
+  'QUESTION_ANSWERED': 2
+}
