+++ lambda/utils/logger.py
@@ -1,0 +1,14 @@
+# encoding: utf-8
+import logging
+
+def init_logger():
+    log_formatter = logging.Formatter("%(asctime)s - %(name)s - %(levelname)s - %(message)s")
+    root_logger = logging.getLogger()
+    root_logger.setLevel(logging.DEBUG)
+
+    console_handler = logging.StreamHandler()
+    console_handler.setFormatter(log_formatter)
+    root_logger.addHandler(console_handler)
+
+def get_logger():
+  return logging.getLogger()
